+++ Runtime/Core/UsercentricsLoggerLevel.cs
@@ -1,0 +1,10 @@
+namespace Chartboost.Core.Usercentrics
+{
+    public enum UsercentricsLoggerLevel
+    {
+        None = 0,
+        Error = 1,
+        Warning = 2,
+        Debug = 3,
+    }
+}